Loads/saves favorites. View description.
So it loads/saves favorites with a few caveats.

1) For some reason it's having trouble saving when you unfavorite an
item when you're viewing a list of questions (will fix this this
afternoon)

2) If you have already favorited an item and you then edit that item (eg
upvoting, adding a reply), these changes are not saved. I'm going to
have to resave the favorites list every time there is an edit. Fixing
this afternoon.

3) It's adding the items to the list in the order GSON is giving it.
This means it's not sorted. I'll have to sort the list by date after
it's loaded.