Lots of changes to load/save (see description)
1) Load/save methods have been implemented for reading list.

2) All saving has been moved to MainActivity's onResume() method. This
makes sense because it will execute whenever the user returns to the
main screen after replying, favoriting a question, etc.

3) I have created a LoadSave static boolean that will indicate whether
there are unsaved changes, and starting to implement it in various
adapters. This doesn't work yet because we don't have unique ID's for
questions or users. So for now the phone will save every single time
onResume executes. See the lengthy comments in onResume for a better
explanation.