GeoLocation complete. See description.
The 'Set Location' activity is complete and works fine on my phone. I
haven't tested it on the virtual machine because I'm not sure how to
emulate GPS results on it yet.

You can edit the city and country fields yourself, or clicking the get
from GPS button will fill them in for you. You then need to click on the
save changes button.

All city and country values are stored in 3 static variables in User.

city = the city
country = the country
use_location = the boolean of whether the user wants to use location
data.

I also have latitude/longitude stored in LocationManager if we need it
(maybe if city and country are set to unknown, which means GPS couldn't
find it).

To do (maybe): It continues to search for a GPS lock until you exit the
get location activity. Maybe add a timeout or allow you to get your
location from network towers.

Saving the city and country in each question's JSON string should allow
us to query for questions near us pretty easily.