simplified fav/readinglist Answer functionality
- Basically, when you click on fav/readinglist on an Answer, it
favourites or adds to reading list the corresponding question.
- There’s no update on those buttons for Answer. It just allows you to
favourite it not unfavorite it. (Answers won’t display whether or not
the question was favorited through that answer or not. It only changes
(filled in or not filled in) when you initially click on it. If you
click on it again & the question is already favorited, it won’t change
fills or anything. (looks like nothing happens - gives a toast message)
- Only way to unfav it is to do so directly though the question.
- This is just a simplified behaviour but it seems reasonable since
logically, favoriting an answer is basically just favouriting the
question anyways